extract mmseqs1 runtime into module
